Fix sign compare warning for gcc
Signed value should not compared with unsigned value,
excpet developer use the force type cast statement,
and make sure it is safe.

Type: Code Improvement

+++ ovxlib/src/ops/vsi_nn_op_squeeze.c
@@ -124,7 +124,7 @@ static vsi_bool op_setup
             {
                 int32_t rank = self->nn_param.squeeze.axis[i];
 
-                rank = rank < 0 ? rank + inputs[0]->attr.dim_num : rank;
+                rank = rank < 0 ? rank + (int32_t)inputs[0]->attr.dim_num : rank;
 
                 if ( !shouldSqueeze[rank] )
                 {
